+++ src/main/java/com/cev/sircapcev/entity/EquipoEntity.java
@@ -12,10 +12,10 @@ import javax.persistence.Table;
 public class EquipoEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String serial;
-    private Long idBodega;
-    private Long idCuenta;
+    private Integer idBodega;
+    private Integer idCuenta;
     private String estado;
     private Date fechaMod;
 
@@ -24,7 +24,7 @@ public class EquipoEntity {
     }
 
 
-    public EquipoEntity(Long id, String serial, Long idBodega, Long idCuenta, String estado, Date fechaMod) {
+    public EquipoEntity(Integer id, String serial, Integer idBodega, Integer idCuenta, String estado, Date fechaMod) {
         this.id = id;
         this.serial = serial;
         this.idBodega = idBodega;
@@ -33,11 +33,11 @@ public class EquipoEntity {
         this.fechaMod = fechaMod;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -49,19 +49,19 @@ public class EquipoEntity {
         this.serial = serial;
     }
 
-    public Long getIdBodega() {
+    public Integer getIdBodega() {
         return this.idBodega;
     }
 
-    public void setIdBodega(Long idBodega) {
+    public void setIdBodega(Integer idBodega) {
         this.idBodega = idBodega;
     }
 
-    public Long getIdCuenta() {
+    public Integer getIdCuenta() {
         return this.idCuenta;
     }
 
-    public void setIdCuenta(Long idCuenta) {
+    public void setIdCuenta(Integer idCuenta) {
         this.idCuenta = idCuenta;
     }
