+++ src/main/java/com/cev/sircapcev/entity/TipoSolicitudEntity.java
@@ -10,7 +10,7 @@ import javax.persistence.Table;
 public class TipoSolicitudEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nombre;
     
 
@@ -18,16 +18,16 @@ public class TipoSolicitudEntity {
     public TipoSolicitudEntity() {
     }
 
-    public TipoSolicitudEntity(Long id, String nombre) {
+    public TipoSolicitudEntity(Integer id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
